requirements.txt
torch>=1.1
torchvision~=0.14.1
numpy~=1.21.5
tqdm~=4.64.1
tensorboard>=1.14

pandas~=1.3.5
pillow~=9.3.0